[pdsl_core] Remove env_logger as dev_dependency
The problem is that env_logger is obviously not no_std compatible.
At first you would think that this is not a problem, since it is only a dev-dependency that is allowed to pull the std lib.
However, due to a Cargo "bug" all dev-dependencies and normal dependencies need to have shared feature flags.
So since env_logger depends on log and pdsl_core depends on log both need to share their feature set which is chosen to be just the one that pulls std - what a mess ...

We now need our own logging facilities that are no_std compatible.