[pdsl_core] Remove memory::HashMap and memory::HashSet
From previous commit:
The problem with memory::Hash{Map,Set} is that it is not compatible with no_std environments.
It requires random state which is currently not a thing in the SRML and also it might use floats internally (e.g. for fill rate) which are also currently disabled due to non-determinism.